pyproject.toml
[tool.poetry]
name = "ascender"
version = "0.1.0"
description = "Project template for XCCV group of cvpaper.challenge."
license = "MIT"
authors = ["cvpaper.challenge XCCV group"]
maintainers = ["yoshihiro fukuhara <gatheluck@gmail.com>"]
readme = "README.md"
keywords = [
    "ascender",
    "Python",
    "projects",
    "project templates",
    "project directory",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: GPU :: NVIDIA CUDA",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Natural Language :: Japanese",
    "Programming Language :: Python",
    "Topic :: Software Development",
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
pyannote-audio = "^3.1.1"
soundfile = "^0.12.1"
openai-whisper = "^20231117"
huggingface = "^0.0.1"
ffmpeg = "^1.4"
nltk = "^3.8.1"
huggingface-hub = "^0.30.2"
transformers = "^4.51.3"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.10.1"
mdformat-gfm = "^0.3.5"
mdformat-frontmatter = "^2.0.1"
mdformat_footnote = "^0.1.1"
flake8 = "^5.0.0"
mypy = "^1.0.1"
pytest-cov = "^4.0.0"
huggingface-hub = {extras = ["cli"], version = "^0.30.2"}

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = 3.8
# following setting is same as pysen
# https://github.com/pfnet/pysen/blob/main/setup.cfg#L12
check_untyped_defs = true
disallow_any_decorated = false
disallow_any_generics = false
disallow_any_unimported = false
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = false
disallow_untyped_defs = true
ignore_errors = false
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
strict_equality = true
strict_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
